create the file to hold all the function and import them in index.js
