+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "webpack --mode development",
     "test": "echo \"Error: no test specified\" && exit 1",
-     "start": "webpack serve --open",
+    "start": "webpack serve --open",
     "build": "webpack"
   },
   "repository": {
@@ -24,7 +24,7 @@
     "css-loader": "^6.7.1",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
-    "webpack": "^5.72.1",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.9.1"
   },
